[ui] move DateTimePicker under the community export
