+++ packages/expo-ui/package.json
@@ -31,6 +31,10 @@
       "types": "./build/datetime-picker/index.d.ts",
       "default": "./src/datetime-picker/index.tsx"
     },
+    "./community/datetime-picker": {
+      "types": "./build/community/datetime-picker/index.d.ts",
+      "default": "./src/community/datetime-picker/index.tsx"
+    },
     "./community/bottom-sheet": {
       "types": "./build/community/bottom-sheet/index.d.ts",
       "default": "./src/community/bottom-sheet/index.tsx"
