+++ packages/expo-ui/src/community/datetime-picker/DateTimePicker.tsx
@@ -1,9 +1,9 @@
 import { buildEvent, buildChangeEvent, type DateTimePickerProps } from './types';
-import { DatePicker, type DatePickerProps } from '../swift-ui/DatePicker';
-import { Host } from '../swift-ui/Host';
-import { disabled as disabledModifier, type ModifierConfig, tint } from '../swift-ui/modifiers';
-import { datePickerStyle } from '../swift-ui/modifiers/datePickerStyle';
-import { environment } from '../swift-ui/modifiers/environment';
+import { DatePicker, type DatePickerProps } from '../../swift-ui/DatePicker';
+import { Host } from '../../swift-ui/Host';
+import { disabled as disabledModifier, type ModifierConfig, tint } from '../../swift-ui/modifiers';
+import { datePickerStyle } from '../../swift-ui/modifiers/datePickerStyle';
+import { environment } from '../../swift-ui/modifiers/environment';
 
 type DatePickerStyleType = 'automatic' | 'compact' | 'graphical' | 'wheel';
