+++ packages/expo-ui/src/community/datetime-picker/index.tsx
@@ -1,0 +1,11 @@
+import { DateTimePicker } from './DateTimePicker';
+
+export {
+  type DateTimePickerEvent,
+  type DateTimePickerChangeEvent,
+  type DateTimePickerProps,
+} from './types';
+
+export default DateTimePicker;
+// named export needed for docs generator
+export { DateTimePicker };
